Adjust Camera Detail Layout
