+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/components/CameraConfigure/CameraConfigureInfo.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, FC, useState, useCallback } from 'react';
-import { Flex, Text, Status, Button, Loader, Grid, Alert, Input } from '@fluentui/react-northstar';
+import { Flex, Text, Button, Loader, Grid, Alert } from '@fluentui/react-northstar';
 import { Link, useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
@@ -10,28 +10,16 @@ import {
   thunkGetTrainingMetrics,
   thunkGetInferenceMetrics,
   resetStatus,
-  updateProjectData,
-  thunkUpdateProbThreshold,
 } from '../../store/project/projectActions';
 import { Project, Status as CameraConfigStatus } from '../../store/project/projectTypes';
 import { State } from '../../store/State';
-import { useParts } from '../../hooks/useParts';
 import { useQuery } from '../../hooks/useQuery';
-import { LiveViewContainer } from '../LiveViewContainer';
-import { AOIData } from '../../type';
 
-export const CameraConfigureInfo: React.FC<{ projectId: number; AOIs: AOIData }> = ({ projectId, AOIs }) => {
-  const {
-    error,
-    data: project,
-    trainingLog,
-    status,
-    trainingMetrics,
-    inferenceMetrics,
-    isLoading,
-  } = useSelector<State, Project>((state) => state.project);
+export const CameraConfigureInfo: React.FC<{ projectId: number }> = ({ projectId }) => {
+  const { error, trainingLog, status, trainingMetrics } = useSelector<State, Project>(
+    (state) => state.project,
+  );
   const allTrainingLog = useAllTrainingLog(trainingLog);
-  const parts = useParts();
   const dispatch = useDispatch();
   const cameraName = useQuery().get('name');
   const isDemo = useQuery().get('isDemo') === 'true';
@@ -77,10 +65,6 @@ export const CameraConfigureInfo: React.FC<{ projectId: number; AOIs: AOIData }>
     return () => dispatch(resetStatus());
   }, []);
 
-  const isCameraOnline = [CameraConfigStatus.FinishTraining, CameraConfigStatus.StartInference].includes(
-    status,
-  );
-
   return (
     <Flex column gap="gap.large">
       <h1>Configuration</h1>
@@ -92,68 +76,6 @@ export const CameraConfigureInfo: React.FC<{ projectId: number; AOIs: AOIData }>
         </>
       ) : (
         <>
-          <ListItem title="Status">
-            <CameraStatus online={isCameraOnline} />
-          </ListItem>
-          <ListItem title="Configured for">
-            {parts
-              .filter((e) => project.parts.includes(e.id))
-              .map((e) => e.name)
-              .join(', ')}
-          </ListItem>
-          <Flex column gap="gap.small">
-            <LiveViewContainer showVideo={true} initialAOIData={AOIs} cameraId={project.camera} />
-          </Flex>
-          <ListItem title="Maximum">
-            <Input
-              value={project.probThreshold}
-              onChange={(_, { value }): void => {
-                dispatch(updateProjectData({ probThreshold: value }));
-              }}
-            />
-            <span>%</span>
-            <Button
-              primary
-              content="Update Confidence Level"
-              onClick={(): void => {
-                dispatch(thunkUpdateProbThreshold());
-              }}
-              disabled={!project.probThreshold || isLoading}
-              loading={isLoading}
-            />
-          </ListItem>
-          <Grid columns={2} styles={{ rowGap: '20px' }}>
-            <ListItem title="Success Rate">
-              <Text styles={{ color: 'rgb(244, 152, 40)', fontWeight: 'bold' }} size="large">
-                {`${inferenceMetrics.successRate}%`}
-              </Text>
-            </ListItem>
-            <ListItem title={`Running on ${inferenceMetrics.isGpu ? 'GPU' : 'CPU'} (accelerated)`}>{`${
-              Math.round(inferenceMetrics.averageTime * 100) / 100
-            }/ms`}</ListItem>
-            <ListItem title="Successful Inferences">{inferenceMetrics.successfulInferences}</ListItem>
-          </Grid>
-          <ListItem title="Unidentified Items">
-            <Text styles={{ margin: '5px' }} size="large">
-              {inferenceMetrics.unIdetifiedItems}
-            </Text>
-            <Button
-              content="Identify Manually"
-              primary
-              styles={{
-                backgroundColor: 'red',
-                marginLeft: '100px',
-                ':hover': {
-                  backgroundColor: '#A72037',
-                },
-                ':active': {
-                  backgroundColor: '#8E192E',
-                },
-              }}
-              as={Link}
-              to="/manual"
-            />
-          </ListItem>
           {trainingMetrics.prevConsequence && (
             <>
               <Text>Previous Model Metrics</Text>
@@ -170,12 +92,14 @@ export const CameraConfigureInfo: React.FC<{ projectId: number; AOIs: AOIData }>
             recall={trainingMetrics.curConsequence?.recall}
             mAP={trainingMetrics.curConsequence?.mAP}
           />
-          <Button primary disabled={isDemo} onClick={onDeleteConfigure}>
-            Delete Configuration
-          </Button>
-          <Button primary as={Link} to="/partIdentification">
-            Edit Configuration
-          </Button>
+          <Flex gap="gap.medium">
+            <Button primary as={Link} to="/partIdentification">
+              Edit Configuration
+            </Button>
+            <Button primary disabled={isDemo} onClick={onDeleteConfigure}>
+              Delete Configuration
+            </Button>
+          </Flex>
         </>
       )}
     </Flex>
@@ -227,32 +151,5 @@ const ConsequenceDashboard: FC<ConsequenceDashboardProps> = ({ precision, recall
         </Text>
       </div>
     </Grid>
-  );
-};
-
-const ListItem = ({ title, children }): JSX.Element => {
-  return (
-    <Flex vAlign="center" gap="gap.medium">
-      <Text style={{ width: '200px' }} size="large">{`${title}: `}</Text>
-      {typeof children === 'string' || typeof children === 'number' ? (
-        <Text size="large">{children}</Text>
-      ) : (
-        children
-      )}
-    </Flex>
-  );
-};
-
-const CameraStatus = ({ online }): JSX.Element => {
-  const text = online ? 'Online' : 'Offline';
-  const state = online ? 'success' : 'unknown';
-
-  return (
-    <Flex gap="gap.smaller" vAlign="center">
-      <Status state={state} />
-      <Text styles={{ margin: '5px' }} size="large">
-        {text}
-      </Text>
-    </Flex>
   );
 };
