+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui/src/pages/CameraDetails.tsx
@@ -32,8 +32,8 @@ const CameraDetails: FC = (): JSX.Element => {
 
   return (
     <Grid columns="2" design={{ height: '100%' }}>
-      <CameraDetailInfo id={camera.id} name={name} rtsp={camera.rtsp} />
-      {hasProject ? <CameraConfigureInfo projectId={projectId} AOIs={aoiData} /> : <CreateCameraConfig />}
+      <CameraDetailInfo id={camera.id} name={name} rtsp={camera.rtsp} AOIs={aoiData} />
+      {hasProject ? <CameraConfigureInfo projectId={projectId} /> : <CreateCameraConfig />}
     </Grid>
   );
 };
